pyproject.toml
[project]
name = "vertex_api"
description = ""
version = "1.0.0"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
#
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["google-auth", "requests", "pillow", "numpy"]

[project.urls]
Repository = "https://github.com/linfen0/comfyUI_Vertex_API"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/linfen0/comfyUI_Vertex_API/blob/master/%E4%BB%8E0%E5%BC%80%E5%A7%8B%E8%8E%B7%E5%8F%96Vertex%20AI%20%E5%87%AD%E8%AF%81%E7%9A%84%E6%96%B9%E6%B3%95.md"
"Bug Tracker" = "https://github.com/linfen0/comfyUI_Vertex_API/issues"

[tool.comfy]
PublisherId = "rinflan"
DisplayName = "comfyUI_Vertex_API"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

